common-components/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "common-components"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "bnachlieli", email = "bar2133@gmail.com" }]
requires-python = ">=3.12.9"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic-settings",
    "pyyaml",
    "redis",
    "python-multipart",
    "pyjwt",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "pytest-cov", "flake8", "mypy"]
